catch IndexError

+++ yeti_skate/views.py
@@ -209,9 +209,12 @@ class DeleteYetiSkateSessionView(LoginRequiredMixin, DeleteView):
 
         # Clear session from the cart
         skate_date = self.model.objects.filter(id=kwargs['pk']).values_list('skate_date', flat=True)
-        cart_date = self.skate_date_model.objects.filter(id=skate_date[0])
-        # print(cart_date[0])
-        cart_item = Cart.objects.filter(item=Program.objects.all().get(id=7).program_name, event_date=cart_date[0].skate_date).delete()
+        if len(skate_date) == 0:
+           return redirect('yeti_skate:yeti-skate')
+        else:
+            cart_date = self.skate_date_model.objects.filter(id=skate_date[0])
+            # print(cart_date[0])
+            cart_item = Cart.objects.filter(item=Program.objects.all().get(id=7).program_name, event_date=cart_date[0].skate_date).delete()
 
         # Set success message and return
         messages.add_message(self.request, messages.SUCCESS, 'You have been removed from that skate session!')
